DOTO-2 Projects
DOTO-2 Project is the final project for the Human and Computer Interaction course. This project is my (@alyamelati) collaborating project with @AzizahNabilahAnwar and @BabyAyuAlicya).

DOTO-2 is a fighting game that can be played individually or in groups and is a sequel or continuation of the game Doto 1. On the website, users can register an account, watch tutorials on how to play, character descriptions, and so on.